pyproject.toml
[tool.poetry]
name = "browser-automation-launcher"
version = "0.1.0"
description = "Browser automation launcher with SQS integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.0"
aioboto3 = "^13.0.0"
aiohttp = "^3.9.0"
pydantic = "^2.9.0"
psutil = "^6.0.0"
python-dotenv = "^1.1.1"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.0"
pre-commit = "^4.3.0"
selenium = "^4.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
